refactor(backend): SF Object 정합 P0 hotfix — 데이터 손실/오버플로우 위험 6건 즉시 정합
36개 sobject 정합 분석에서 발견된 P0 위험 항목 즉시 적용:

- Claim.defectQuantity / purchaseAmount: Int → Long (SF double 18,0 오버플로우 위험)
- Claim.productCode: 20 → 1300 (SF string 1300)
- Claim.defectDescription: 1000 → 4000 (SF textarea 4000)
- UploadFile.recordId: 18 → 40, size: 50 → 100
- AccountCategoryMaster.accountCode: 20 → 255
- MonthlySalesHistory.externalKey: 30 → 40
- Promotion.promotionNumber: 20 → 80

Flyway V99__sf_align_p0_hotfix.sql 동봉. DTO/Service/Test Long 정합 포함.
근거: docs/plan/old_source_260408/sf-object-meta/README.md §6.6

+++ backend/src/main/kotlin/com/otoki/powersales/claim/dto/response/ClaimListItemResponse.kt
@@ -12,7 +12,7 @@ data class ClaimListItemResponse(
     val categoryLabel: String?,
     val subcategoryValue: String?,
     val subcategoryLabel: String?,
-    val defectQuantity: Int?,
+    val defectQuantity: Long?,
     val status: String,
     val statusLabel: String,
     val createdAt: LocalDateTime
